+++ x/staking/keeper/unbonding.go
@@ -200,9 +200,7 @@ func (k Keeper) SetValidatorByUnbondingID(ctx context.Context, val types.Validat
 		return err
 	}
 
-	valKey := types.GetValidatorKey(valAddr)
-	valKey = valKey[2:] // leave prefix bytes and address length
-	if err = k.UnbondingIndex.Set(ctx, id, valKey); err != nil {
+	if err = k.UnbondingIndex.Set(ctx, id, valAddr); err != nil {
 		return err
 	}